chore: rename hook directory for consistency
