+++ packages/react/src/utilities/index.ts
@@ -1,11 +1,11 @@
 // Public utilities
-export * from './autoId';
 export * from './children';
 export * from './environment';
 export * from './externalClick';
 export * from './forwardedRef';
-export * from './link';
+export * from './id';
 export * from './isomorphicLayoutEffect';
+export * from './link';
 export * from './mediaQuery';
 export * from './popper';
 export * from './popperTriggers';
